+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- :fire: major improvements to `AppFactory` and `AppFactoryHook`s :fire:
+- :fire: major improvements to `AppFactory` and `AppFactoryHook` :fire:
     - support using a custom subclass of `FlaskUnchained` using `AppFactory.APP_CLASS`
     - support using a custom subclass of `AppFactory`
     - support passing all kwargs to `Flask` by setting the same names upper-cased in `unchained_config`
@@ -104,7 +104,10 @@ class YourBundle(AppBundle):
 ### Internals
 
 - move `Bundle` and `AppBundle` into the `flask_unchained.bundles` module
-- move `BundleBlueprint` into the `flask_unchained.bundles._blueprint` module
+- move `BundleBlueprint` into the `flask_unchained.bundles.controler.bundle_blueprint` module
+- move `_DeferredBundleFunctions` into `flask_unchained.unchained`, rename it to `DeferredBundleFunctions`
+- make `_has_views`, `_blueprint_names`, `_static_folders`, `is_top_bundle` and `_has_hierarchy_name_conflicts` methods on `Bundle` properties
+- rename double-negative `reverse_mro` parameter for `Bundle._iter_class_hierarchy` to `mro`
 - warn when identical routes are registered
 - make `AppFactory` methods `load_bundles`, `load_bundle` and `is_bundle` classmethods
 - add a noop `ViewsHook` to consolidate logic for defining and loading `views_module_names`